Add snapshot tests with Playwright for custom-payment-flow
I heard from Paul and Hide that Playwright deals with iframes well and
we are trying to see if it's great for testing the sample apps. As a
touchstone, I added snapshot tests for custom-payment-flow under the
`playwright/` directory.

```
playwright/
├── custom-payment-flow-e2e-html.spec.ts # tests for custom-payment-flow/client/html
├── custom-payment-flow-e2e-html.spec.ts-snapshots # snapshots (screenshots) for the test above
├── custom-payment-flow-e2e-react-cra.spec.ts # tests for custom-payment-flow/client/react-cra
├── custom-payment-flow-e2e-react-cra.spec.ts-snapshots # snapshots (screenshots) for the test above
├── node_modules
├── package.json
└── package-lock.json
```

The tests run in the job `snapshot_test` in the `ci.yml`. Each test
just takes a screenshot of each payment page and compares it with
the one taken in advance (saved under *-snapshots/ dir).  If there
are differences between the two images, the test fails.

* Successful example: https://github.com/hibariya/accept-a-payment/actions/runs/3327818197/jobs/5503086362#step:5:434
* Failure example: https://github.com/hibariya/accept-a-payment/actions/runs/3327875035/jobs/5503212013#step:5:557

If some of the tests fails, we can inspect the following images as
the artifact of the CI run
(like [this](https://github.com/hibariya/accept-a-payment/actions/runs/3327875035#artifacts)).

* The expected screenshot
* The actual screenshot
* The visual diff

See Also:

* http://cuketest.com/playwright/docs/test-snapshots/